+++ tower-http/CHANGELOG.md
@@ -11,7 +11,12 @@ None.
 
 ## Breaking changes
 
-None.
+- Change the response body error type of `Compression` and `Decompression` to
+  `Box<dyn std::error::Error + Send + Sync>`. This makes them usable if the body
+  they're wrapping uses `Box<dyn std::error::Error + Send + Sync>` as its error
+  type which they previously weren't.
+- Remove `BodyOrIoError`. Its been replaced with `Box<dyn std::error::Error +
+  Send + Sync>`.
 
 # 0.1.0 (May 27, 2021)